rename BranchesTable
