+++ src/BranchesTable.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-let DashboardRow = require('./DashboardRow');
+let DashboardRow = require('./DashboardRow.jsx');
 
 let BranchesTable = React.createClass({
 
@@ -10,14 +10,14 @@ let BranchesTable = React.createClass({
 
     render: function() {
         return (
-            <table class="table table-striped table-bordered table-condensed">
+            <table className="table table-striped table-bordered table-condensed">
                 <thead>
                 <tr>
                     <th>Git branch</th>
                     <th>GitHub</th>
                 </tr>
                 </thead>
-                <tbody id="branchesTable">
+                <tbody>
                 {
                     this.props.branches.map(branch => {
                         return <DashboardRow branch={branch} key={branch}/>;
